// src/models/messages/message.interface.ts
import { Profile } from '../profile/profile.interface';

export interface Message {
    userFromId: string;

    userFromProfile: {
        firstName: string;
        lastName: string;
    }

    userToId: string;

    userToProfile: {
        firstName: string;
        lastName: string;
    }

    content: string;
    uploadImage: string;
    time: string;
    durum: string;

}